--- src/kili/services/json_response/__init__.py
@@ -1,4 +1,0 @@
-"""
-This module contains an alpha feature for parsing label json response.
-Note that this feature is still under development and is not yet suitable for use by general users.
-"""
