--- src/kili/services/json_response/annotation.py
@@ -1,226 +1,0 @@
-# Feature is still under development and is not yet suitable for use by general users.
-"""Module for the "annotations" key parsing of a job response."""
-
-import functools
-from typing import Dict
-
-from typeguard import typechecked
-
-from .category import Category, CategoryList
-from .decorators import for_all_properties
-from .exceptions import AttributeNotCompatibleWithJobError
-
-
-class _BaseAnnotation:
-    """Class for parsing the "annotations" key of a job response.
-
-    It is used as a base class for the common properties of all types of annotations.
-    """
-
-    def __init__(self, annotation_json: Dict, job_interface: Dict) -> None:
-        """Class for Annotation parsing.
-
-        Args:
-            annotation_json: Dict of an annotation. It is the value of
-                the "annotations" key of a job response.
-            job_interface: Job interface of the job.
-        """
-        self.json_data: Dict = annotation_json
-        self.job_interface: Dict = job_interface
-
-        self.is_required_job = job_interface["required"]
-
-    def _cast_categories(self):
-        """Casts the categories list of the job payload to CategoryList object."""
-        if "categories" not in self.json_data:
-            return
-
-        if not isinstance(self.json_data["categories"], CategoryList):
-            self.json_data["categories"] = CategoryList(
-                self.job_interface, self.json_data["categories"]
-            )
-
-    @property
-    def categories(self) -> CategoryList:
-        """Returns the list of categories of the annotation."""
-        self._cast_categories()
-        return self.json_data["categories"]
-
-    @property
-    def category(self) -> Category:
-        """Returns the category of the annotation if there is only one category.
-
-        Else raises an error.
-        """
-        if self.job_interface["content"]["input"] not in ("radio", "singleDropdown"):
-            raise AttributeNotCompatibleWithJobError("category")
-
-        if "categories" not in self.json_data and not self.is_required_job:
-            return None  # type: ignore
-
-        self._cast_categories()
-        return self.json_data["categories"][0]
-
-    @property
-    def mid(self) -> str:
-        """Returns the mid of the annotation."""
-        return self.json_data["mid"]
-
-    @mid.setter
-    @typechecked
-    def mid(self, mid: str) -> None:
-        """Sets the mid of the annotation."""
-        if len(mid) == 0:
-            raise ValueError("mid must be non-empty.")
-        self.json_data["mid"] = mid
-
-    @property
-    def children(self):
-        """Not implemented yet."""
-
-
-class EntityAnnotation(_BaseAnnotation):
-    """Class for parsing the "annotations" key of a job response for named entities recognition."""
-
-    @property
-    def begin_offset(self) -> int:
-        """Returns the begin offset of the annotation."""
-        return self.json_data["beginOffset"]
-
-    @begin_offset.setter
-    @typechecked
-    def begin_offset(self, begin_offset: int) -> None:
-        """Sets the begin offset of the annotation."""
-        if begin_offset < 0:
-            raise ValueError(f"begin_offset must be positive, got {begin_offset}")
-        self.json_data["beginOffset"] = begin_offset
-
-    @property
-    def end_offset(self) -> int:
-        """Returns the end offset of the annotation."""
-        return self.json_data["endOffset"]
-
-    @end_offset.setter
-    @typechecked
-    def end_offset(self, end_offset: int) -> None:
-        """Sets the end offset of the annotation."""
-        if end_offset < 0:
-            raise ValueError(f"end_offset must be positive, got {end_offset}")
-        self.json_data["endOffset"] = end_offset
-
-    @property
-    def content(self) -> str:
-        """Returns the content of the annotation."""
-        return self.json_data["content"]
-
-    @content.setter
-    @typechecked
-    def content(self, content: str) -> None:
-        """Sets the content of the annotation."""
-        self.json_data["content"] = content
-
-
-class _BaseAnnotationWithTool(_BaseAnnotation):
-    """Base class for annotations with a "type" key (tool used to create the annotation)."""
-
-    @property
-    def type(self) -> str:
-        """Returns the tool of the annotation.
-
-        One of: "rectangle", "polygon", or "semantic".
-        """
-        return self.json_data["type"]
-
-
-class PointAnnotation(_BaseAnnotationWithTool):
-    """Class for parsing the "annotations" key of a job response for 1D object detection jobs."""
-
-    @property
-    def point(self):
-        """Not implemented yet."""
-
-
-class _Base2DAnnotation(_BaseAnnotationWithTool):
-    """Base class for 2D annotations."""
-
-    @property
-    def bounding_polygon(self):
-        """Returns the polygon of the object contour."""
-        return self.json_data["boundingPoly"]
-
-    @property
-    def score(self):
-        """Returns the score which is the confidence of the object detection.
-
-        Useful when a pre-annotation model is used.
-        """
-        return self.json_data["score"]
-
-
-class BoundingPolygonAnnotation(_Base2DAnnotation):
-    """Class for parsing the "annotations" key of a job response for 2D object detection jobs."""
-
-
-class VideoAnnotation(_Base2DAnnotation):
-    """Class for parsing the "annotations" key of a job response for video object detection jobs."""
-
-    @property
-    def is_key_frame(self) -> bool:
-        """Returns a boolean indicating if the timestamp or frame is used for interpolation."""
-        return self.json_data["isKeyFrame"]
-
-
-class PoseEstimationAnnotation(_BaseAnnotationWithTool):
-    """Class for parsing the "annotations" key of a job response for pose estimation jobs."""
-
-    @property
-    def kind(self):
-        """Not implemented yet."""
-
-    @property
-    def points(self):
-        """Not implemented yet."""
-
-
-ANNOTATIONS_VALID_ATTRIBUTES_FOR_ML_TASK = {
-    "NAMED_ENTITIES_RECOGNITION": ("begin_offset", "mid", "categories", "category", "content"),
-    "OBJECT_DETECTION": ("mid", "categories", "category", "type"),
-    "POSE_ESTIMATION": ("mid", "categories", "category", "type"),
-}
-
-
-def is_attribute_compatible_with_mltask(attribute_name: str, job_interface: Dict) -> bool:
-    """Returns True if the attribute is compatible with the job interface mlTask."""
-    return attribute_name in ANNOTATIONS_VALID_ATTRIBUTES_FOR_ML_TASK.get(
-        job_interface["mlTask"], tuple()
-    )
-
-
-def check_attribute_compatible_with_mltask(func):
-    """Raises an error if the decorated method is not compatible with the job interface mlTask."""
-
-    @functools.wraps(func)
-    def wrapper(self, *args, **kwargs):
-        attribute_name = func.__name__
-        if not is_attribute_compatible_with_mltask(attribute_name, self.job_interface):
-            raise AttributeNotCompatibleWithJobError(attribute_name)
-        return func(self, *args, **kwargs)
-
-    return wrapper
-
-
-@for_all_properties(check_attribute_compatible_with_mltask)
-# pylint: disable=too-many-ancestors
-class Annotation(
-    EntityAnnotation,
-    PointAnnotation,
-    BoundingPolygonAnnotation,
-    VideoAnnotation,
-    PoseEstimationAnnotation,
-):
-    """Class for parsing the "annotations" key of a job response.
-
-    Contains all attributes that can be found in a job response.
-    """
-
-    # empty class: all properties should be inherited from the base classes
