+++ tests/services/label_data_parsing/creation/test_resp_creation.py
@@ -1,4 +1,4 @@
-from kili.services.json_response.json_response import ParsedJobs
+from kili.services.label_data_parsing.json_response import ParsedJobs
 
 
 def test_create_transcription_label():
