+++ tests/services/label_data_parsing/parsing/test_resp_parsing_children.py
@@ -1,6 +1,6 @@
 import pytest
 
-from kili.services.json_response.json_response import ParsedJobs
+from kili.services.label_data_parsing.json_response import ParsedJobs
 
 
 @pytest.mark.skip("Not implemented yet")
